+++ src/GitLabApiClient/IssuesClient.cs
@@ -164,7 +164,7 @@ namespace GitLabApiClient
         /// <param name="projectId">The ID, path or <see cref="Project"/> of the project.</param>
         /// <param name="issueIid">The IID of an issue.</param>
         /// <param name="request">Create issue note request.</param>
-        public async Task<Note> CreateNoteAsync(ProjectId projectId, int issueIid, CreateIssueNoteRequest request) =>
+        public async Task<Note> CreateNoteAsync(ProjectId projectId, long issueIid, CreateIssueNoteRequest request) =>
             await _httpFacade.Post<Note>($"projects/{projectId}/issues/{issueIid}/notes", request);
 
         /// <summary>
